Move container-agnostic watcher producer helpers out of _k8s_common
The producer state, the XCom-backup callbacks, the outlet-URI map and the
retry-aware execute shared by the Kubernetes and GKE watcher producers do
not depend on Kubernetes, but lived in `_k8s_common`, which imports
`kubernetes.client` and the cncf provider at module level. A producer for
another runtime could not import them without the Kubernetes stack.

Move them, verbatim apart from docstring wording, to
`cosmos/operators/_watcher/producer.py` together with the context-holder
keys. `_k8s_common` keeps the pod callback, its injection and the pod
manager builder, re-exports the moved names, and its
`init_watcher_producer` becomes the K8s flavour: shared init plus callback
injection. `watcher_kubernetes.py` and `watcher_gcp_gke.py` are unchanged.

Tests that patched the XCom helpers on `_k8s_common` now patch them on
the module that calls them.

Prepares #2988.
